refact : 사용하지 않는 API 제거 및 컨트롤러 이동
